* errno:
  Replacement of the declaration for the variable "errno" by an include

* braces:
  Clarification of an if instruction

* string:
  Inclusion of <string.h>

* binary_search:
  Addition of a type compatible comparison function for a binary search

* temporary_files:
  Replacement of the legacy function "mktemp" by the safer function "mkstemp"

* make_functions_static:
  Scope reduction for some functions by the mark "static"

+++ mstring.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <stdarg.h>
 #include <ctype.h>
+#include <string.h>
 #include "mstring.h"
 
 /* parameters about string length.  START is the starting size and
@@ -13,7 +14,7 @@ void msprintf(struct mstring *s, const char *fmt, ...)
 {
 static char	buf[4096];	/* a big static buffer */
 va_list		args;
-int		len;
+size_t		len;
 
     if (!s || !s->base) return;
     va_start(args, fmt);
